feat: dependency_container
This is the first work to automate the creation of dependency
containers. The idea is to eventually have the same output as the [manual
implementation outlined in my blog](https://chesedo.me/blog/manual-dependency-injection-rust/).

This first implementation just auto wires the dependencies correctly.
Ie. it created the public method as outlined in the blog post. Async
dependencies are not supported with this commit. Nor are lifetimes.