feat(di): handle nested `impl Trait`s correctly
Sometimes an `impl Trait` can be more than one level of generics deep.
Or it might come from more than just the immediate children. This PR
handles these cases correctly.

It also expands the child dependency explicitly rather than calling a
function to get the child dependencies. This is needed to prevent
premature casting to `impl` types which leads to a compile error.